Split system builtins
