+++ src/builtins.h
@@ -37,6 +37,8 @@ int builtin_exec(char **args);
 int builtin_command(char **args);
 int builtin_time(char **args);
 int builtin_times(char **args);
+int builtin_umask(char **args);
+int builtin_ulimit(char **args);
 int builtin_source(char **args);
 const char **get_builtin_names(void);
 const char *get_alias(const char *name);
